iyes_bevy_extras
Module cleanup

Functions

despawn_all

Convenience system for despawning all entities that match a given query filter

despawn_all_recursive

init_resource

Convenience system for initting a resource of the given type

remove_from_all

Convenience system for removing a component from all entities that match a given query filter

remove_resource

Convenience system for removing a resource of the given type