iyes_bevy_extras::cleanup
Function despawn_all

pub fn despawn_all<F: QueryFilter>(
    world: &mut World,
    query: &mut QueryState<Entity, F>,
)

Convenience system for despawning all entities that match a given query filter

This is useful as an “exit” system in your app states, to clean up large swaths of entities on state transition. For example, you could create a marker component for all of your gameplay entities, and use this system to easily despawn all of them when going back to the main menu.

Consider using [despawn_all_with_recursive][bevy_hierarchy::despawn_all_with_recursive] instead, to ensure you are not left with broken hierarchies. This could happen if you have an entity with the component in a hierarchy where not all entities have the component. This system will only despawn the entities with the component.